Moved Mercure and RabbitMq event listeners to their own subnamespaces
